big refactor.
